+++ MyWhisper/Vocabulary/VocabularyEntry.swift
@@ -1,6 +1,7 @@
 import Foundation
 
-struct VocabularyEntry: Codable, Equatable {
+struct VocabularyEntry: Codable, Equatable, Identifiable {
+    var id: UUID = UUID()
     var wrong: String
     var correct: String
 }